Do not require administer blocks permission to access creation form
The current version of implements hook_entity_create_access to allow
users to create specific types of block content. However the route to
the block content creation form still requires the “administer blocks”
permission. Adding this permission seems to go against the entire 
purpose of this module as it will grant access to all block content
types anyhow.

This is an attempt at fixing this by changing the access control for 
the block content create form route to allow either “administer blocks”
or the new per block content type permissions.